fix: sidecar crate to build fixtures

fix: address feedback


fix: restore common use for benches


fix: address feedback


fix: bad rebase


fix: remove unused dep
