+++ crates/wash-runtime/Cargo.toml
@@ -143,11 +143,9 @@ url = { workspace = true }
 webpki-roots = { workspace = true }
 
 [build-dependencies]
-anyhow = { workspace = true, default-features = true }
+anyhow = { workspace = true }
 tonic-prost-build = { workspace = true, default-features = true }
 pbjson-build = { workspace = true, default-features = true }
-wasi-preview1-component-adapter-provider = "43"
-wit-component = { workspace = true, default-features = false }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["full"] }
@@ -168,6 +166,7 @@ criterion = { version = "0.5", features = ["async_tokio", "html_reports"] }
 # enforces equality between this crate version and the runner binary at
 # run time.
 gungraun = "0.19.1"
+common = { package = "wash-runtime-test-helper", path = "./tests/common/" }
 
 [[bench]]
 name = "http_invoke"
